requirements.txt
streamlit-extras
faker